+++ CHANGELOG.md
@@ -19,3 +19,4 @@
 
 - Hide youtube Shorts everywhere
 - Toggle for hide Shorts in Search results
+- Changed extension name